+++ resolwe_bio/tests/processes/test_support_processors.py
@@ -726,3 +726,27 @@ re-save-file report "${NAME}".txt
             'hs_gatk_wgs_metrics.txt',
             file_filter=filter_comment_lines
         )
+
+    @tag_process('rrbs-metrics')
+    def test_rrbs_metrics(self):
+        with self.preparation_stage():
+            bam = self.run_process('upload-bam', {
+                'src': 'bamclipper/output/TP53.primerclipped.bam',
+                'species': 'Homo sapiens',
+                'build': 'hg19',
+            })
+            genome = self.run_process('upload-genome', {
+                'src': 'bqsr/input/hs_b37_chr17_upto_TP53.fasta.gz',
+                'species': 'Homo sapiens',
+                'build': 'hg19'
+            })
+
+        rrbs_metrics = self.run_process('rrbs-metrics', {
+            'bam': bam.id,
+            'genome': genome.id,
+        })
+
+        self.assertFileExists(
+            rrbs_metrics,
+            'report',
+        )